move all test cases to a common folder
