+++ test/test-integration/src/test/java/org/wso2am/micro/gw/tests/testCases/jwtGenerator/JwtTransformerTestCase.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2am.micro.gw.tests.jwtGenerator;
+package org.wso2am.micro.gw.tests.testCases.jwtGenerator;
 
 import io.netty.handler.codec.http.HttpHeaderNames;
 import org.json.JSONObject;
@@ -24,7 +24,6 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.wso2am.micro.gw.tests.util.HttpResponse;
 import org.wso2am.micro.gw.tests.util.HttpsClientRequest;
-import org.wso2am.micro.gw.tests.util.TestConstant;
 
 import java.util.Base64;
 import java.util.HashMap;
